Implement the model that makes __VA_OPT__ transparent to # and ##.
That is, think of variadic macros that are written with va-opt as having two sets of replacement tokens - one with the contents of all the __VA_OPT__ s, and another where they are replaced by __VA_ARGS__ (i.e. behave like placemarker tokens).
When defining the macro, the macro has to be well defined for both versions.

Think of the choice of which set of (notional) replacement tokens to use as occurring prior to stringization and concatenation and argument substitution and expansion.